Initial requirements for offline "queue for download" functionality.
Use the "App Preferences" table to store whether or not an app has been
queued for download by the user. This table is for info about apps that
is specified by users. Previously it was only "Ignore updates" info.
However it seems that this is similar in nature.

Also reporposed the "Uninstall" button to also be "Download later" (and
"Don't download later" - cancel pending download) button.

The actual downloading of apps when connectivity resumes is not yet
implemented.