demote lint ImpliedQuantity to warning
This is hitting a lot of plurals, causing the CI builds to fail.  But
unfortunately, its getting it wrong.  Its actually for a format in a string
like "Updated today" in languages like pt that need special cases for 0/1
items.  In this case, it makes no sense to say "Updated 0 days ago".