Use app name and version when bluetoothing to peer.
Previously, it would either send "base.apk" (in earlier versions of
F-Droid when bluetoothing an apk from the /data/app/... dirs), or
"install-[random].apk" (if recent F-Droid when copying file to a safe
place to expose via a FileProvider.

This now writes the file to, e.g. "F-Droid-0.103.1.apk" before sending.

Note that this means files are more likely to be overridden when being
sent, if the same apk from two different repositories is either:
 * Sent via bluetooth
 * Prompted to install via the system package manager

However this should still never let malicious people write to that
folder.