remove translation auto-detect hack, it is too hacky
The hack that goes through and checks whether a language is present in
the APK seems to cause random strings to switch to English when the
app is running.  So this removes that hack, and instead switches the
Languages menu to a hard-coded list.  Languages that are not present
or close to complete were removed from the old list.

closes #943
closes #1010