Simplify logic in UpdatesAdapter.
No longer do we try to nicely maintain the state of the adapter in "Updates"
in order to notify the recycler view about changes to its underlying data.
Instead, we just rebuild the entire structure each time a new thing needs
to be shown/removed.

This means no more smooth scrolling to the relevant item after it is
changed, but it results in a far less buggy interface.