Simplify and clean up updates view logic.
Same as how AppDetails2 was recently cleaned up to depend more on
AppUpdateStatusManager.

In addition, it also removes items from the "X apps can be updated"
lower part of the "Updates" view when they are present in the upper half
(i.e. the half showing feedback about the current download/install
progress).