Add final artwork and correctly read colours.
Carrie specified colours earlier, and they were added to the code.
However they were not being read correctly. This changes that so that
lowercase resource names (e.g. "category_games") are used instead.

It also adds the final category artwork, for "Games" which was
missed prior.

The rest still generate colours and patterns if they don't have a colour
or an image specified.