Refactor `ApkProvider.getApkFromAnyRepo` to `getSuggestedApk(Apk)`.
This is really the intention of the method, given it used to accept
a version code and a package name. Now it optionally accepts a sig
also. If present, it will restrict the query to apks with that sig.

Also added to the multi-sig tests to ensure this method takes it into
consideration.