Further refactor the AppListItemController for readability and predictability.
This breaks out subclasses for each specific type of app list item,
allowing for code reuse, but also letting the specific business logic
belonging to each different app list item to be separate.

This is particularly helpful in the following situation:

 * In the search results, it is great to be able to render "App
   downloaded, ready to install" in the same manner as the update tab.

 * In the installed app list, this is not desired. Indeed, the status
   text which should be shown should reference the currently installed
   version and whether the user has ignored any updates.

By separating the AppListItemController into subclasses, it reduced the
need to handle several different types of text view (e.g.
"installedStatus", "status", "ignoredStatus", "downloadReady"), and
replace them all with a "status" and "secondaryStatus" TextView. What is
displayed in status and secondaryStatus is up to the individual
subclasses of AppListItemController.